Added existence check of .vim

+++ setup.sh
@@ -21,9 +21,12 @@ echo "Setting up development environment"
 echo "Installing vim"
 sudo apt-get install vim
 
-echo "Backing up ~/.vim to ~/.vim.bak"
-mv ~/.vim ~/.vim.bak
-mkdir -p ~/.vim/bundle
+if [ -e "~/.vim" ]; then
+    echo "Backing up ~/.vim to ~/.vim.bak"
+    mv ~/.vim ~/.vim.bak
+    mkdir -p ~/.vim/bundle
+fi
+
 
 echo "Installing Vundle"
 git clone https://github.com/gmarik/Vundle.vim ~/.vim/bundle/Vundle.vim
